requirements.txt
Django==5.2.6
djangorestframework==3.16.1
python-dotenv==1.1.1
pillow==11.3.0
psycopg2==2.9.10
requests==2.32.5
djangorestframework_simplejwt==5.5.1
django-cors-headers==4.9.0
django-filter==25.1